+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add spec YAML syntax check to lint target
 - Add new bundle configuration for Apache HTTP Components Proxy Configuration shinesolutions/aem-aws-stack-builder#235
+- Add new  feature to retrieve CRX Package Manager Servlet service state shinesolutions/aem-aws-stack-builder#214
 
 ### Changed
 - Update swagger_aem to 2.4.0